# Conflicts:
#	controllers/message.py
#	store/models/models.py

+++ main_menu/menu.py
@@ -75,6 +75,22 @@ class MainMenuScreen(BaseScreen):
                 self.manager_screen.current = meta.SCREENS.LOGIN_SCREEN
             self.move_menu_backdrop()
 
+    def go_to_subscribers_screen(self):
+        if not self.is_backdrop_front_open:
+            if self.manager_screen.current != meta.SCREENS.LOGIN_SCREEN:
+                self.manager_screen.current = meta.SCREENS.SUBSCRIBERS_SCREEN
+            else:
+                self.manager_screen.current = meta.SCREENS.LOGIN_SCREEN
+            self.move_menu_backdrop()
+
+    def go_to_subscriptions_screen(self):
+        if not self.is_backdrop_front_open:
+            if self.manager_screen.current != meta.SCREENS.LOGIN_SCREEN:
+                self.manager_screen.current = meta.SCREENS.SUBSCRIPTIONS_SCREEN
+            else:
+                self.manager_screen.current = meta.SCREENS.LOGIN_SCREEN
+            self.move_menu_backdrop()
+
     def open_menu_backdrop(self):
         if not self.is_backdrop_front_open:
             if self.manager_screen.current != meta.SCREENS.LOGIN_SCREEN:
@@ -84,9 +100,11 @@ class MainMenuScreen(BaseScreen):
         self.backdrop.open()
 
     def log_out(self):
-        try:
-            self.manager_screen.current = meta.SCREENS.LOGIN_SCREEN
-            AUTHORIZATION.TOKEN = ''
-            os.remove('saved\\access_token')
-        except:
-            pass
+        if not self.is_backdrop_front_open:
+            try:
+                self.manager_screen.current = meta.SCREENS.LOGIN_SCREEN
+                self.move_menu_backdrop()
+                AUTHORIZATION.TOKEN = ''
+                os.remove('saved\\access_token')
+            except:
+                pass
